+++ src/main/java/com/besson/arknights/block/lifecycle/SuckBeverages.java
@@ -1,0 +1,21 @@
+package com.besson.arknights.block.lifecycle;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
+
+public class SuckBeverages extends FurnitureHorizontalFacingBlock {
+    public SuckBeverages(Properties settings) {
+        super(settings);
+    }
+    private static final VoxelShape SHAPE = Block.box(6, 0, 6, 10, 7, 10);
+
+    @Override
+    public VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext) {
+        return SHAPE;
+    }
+}